Rename generators folder to contributors_delegates folder
