+++ lib/src/contributors_delegates/in_place/in_place_union_delegate.dart
@@ -8,8 +8,8 @@ import 'package:data_class_plugin/src/annotations/constants.dart';
 import 'package:data_class_plugin/src/annotations/union_internal.dart';
 import 'package:data_class_plugin/src/common/code_writer.dart';
 import 'package:data_class_plugin/src/contributors/generators/generators.dart';
+import 'package:data_class_plugin/src/contributors_delegates/class_generation_delegate.dart';
 import 'package:data_class_plugin/src/extensions/extensions.dart';
-import 'package:data_class_plugin/src/generators/class_generation_delegate.dart';
 import 'package:data_class_plugin/src/options/data_class_plugin_options.dart';
 import 'package:data_class_plugin/src/visitors/class_visitor.dart';
 import 'package:data_class_plugin/src/visitors/redirected_constructor_visitor.dart';
